So some things are going wrong.
I think we need to decouple the Roku app from the server app. Not sure why that(the server port) is firing up with then the Roku app is launched.

Currently, when I launche the Roku app on its own, it cannot reach the local host 8080.

Then, there is the client app. For some reason, that get’s the line : gRPC Server started, listening on address: *, port: 9090 when it’s up and running. Weird.